+++ src/main/java/courgette/runtime/junit/CourgetteJUnitRunner.java
@@ -4,10 +4,9 @@ import courgette.runtime.CourgetteCallbacks;
 import courgette.runtime.CourgetteProperties;
 import courgette.runtime.CourgetteRunResult;
 import courgette.runtime.CourgetteRunnerInfo;
-import courgette.runtime.CourgetteTestErrorException;
 import courgette.runtime.CourgetteTestFailureException;
-import courgette.runtime.RunStatus;
 import io.cucumber.core.gherkin.Feature;
+import io.cucumber.core.gherkin.Pickle;
 import org.junit.runner.Description;
 import org.junit.runner.notification.Failure;
 import org.junit.runner.notification.RunNotifier;
@@ -18,6 +17,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 public abstract class CourgetteJUnitRunner extends ParentRunner<Feature> {
@@ -80,23 +80,28 @@ public abstract class CourgetteJUnitRunner extends ParentRunner<Feature> {
         featureDescriptions.values().forEach(notifier::fireTestFinished);
     }
 
-    protected void notifyTestFailure(RunNotifier notifier, List<CourgetteRunResult> failures, RunStatus runStatus) {
+    protected void notifyTestFailure(RunNotifier notifier, List<CourgetteRunResult> failures) {
         failures.forEach(failure -> {
             Feature feature = failure.getFeature();
             Description description = featureDescriptions.get(feature);
-            notifier.fireTestFailure(new Failure(description, createFailureThrowable(feature, runStatus)));
+            notifier.fireTestFailure(new Failure(description, createFailureThrowable(feature, failure)));
         });
 
         featureDescriptions.keySet().removeAll(failures.stream().map(CourgetteRunResult::getFeature).collect(Collectors.toList()));
     }
 
-    private Throwable createFailureThrowable(Feature feature, RunStatus runStatus) {
-        final String testId = feature.getName().orElse("Test");
+    private Throwable createFailureThrowable(Feature feature, CourgetteRunResult failure) {
+        String testId = feature.getName().orElse("Test");
 
-        if (RunStatus.ERROR.equals(runStatus)) {
-            return new CourgetteTestErrorException(testId + " was not run due to an error.");
-        } else {
-            return new CourgetteTestFailureException(testId + " failed.");
+        if (failure.getLineId() != null) {
+            Optional<Pickle> scenarioName = feature.getPickles()
+                    .stream().filter(t -> t.getLocation().getLine() == failure.getLineId())
+                    .findFirst();
+
+            if (scenarioName.isPresent()) {
+                testId = testId + " - " + scenarioName.get().getName();
+            }
         }
+        return new CourgetteTestFailureException(testId + " failed.");
     }
 }